requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
slack-sdk==3.23.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pgvector==0.2.4
gunicorn==21.2.0 